feat: Add Loki visual regression
Added Loki visual regression tests for both mobile and desktop. In order
to do this, an upgrade to various dependencies was required, such as
Bootstrap, for compatibility.

The upgrade of Bootstrap from v4 to v5 brought with it a few changes in
default spacing etc. which created an unbalance in the two main columns
on the CV page, so the Social section has been moves to the left column.