Remove Lighthouse checks
Removed this check from the pipeline in order to avoid too many false
negatives. The `foo-software/lighthouse-check-action@master` action that
was used was not reliable when compared to simply running Lighthouse in
the browser and would inconsistently fail with different results each
time, resulting in a game of run checks > fail checks > rerun checks >
pass > done.